refactor:Train model 수정

+++ src/main/java/com/example/planergram/model/Train.java
@@ -17,7 +17,7 @@ public class Train {
     @Id
     @Column(name="train_id",nullable = false, unique = true)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int trainId;
+    private Long trainId;
 
     @Column(name = "departure_time")
     private LocalDateTime departureTime;
